Increase test coverage, modularize, improve code style.
