+++ flows/magnitudes.py
@@ -1,15 +1,17 @@
-from typing import Tuple, Any, Optional
-from astropy.table import Table
-import numpy as np
-from bottleneck import nansum
-from astropy.stats import sigma_clip
-from astropy.modeling import models, fitting
-import matplotlib.pyplot as plt
+from typing import Tuple, Optional
 
-from .target import Target
-from .zeropoint import sigma_from_Chauvenet, bootstrap_outlier
+import matplotlib.pyplot as plt
+import numpy as np
+from astropy.modeling import models, fitting
+from astropy.stats import sigma_clip
+from astropy.table import Table
+from bottleneck import nansum
+
 from .filters import get_reference_filter
+from .target import Target
 from .utilities import create_logger
+from .zeropoint import sigma_from_Chauvenet, bootstrap_outlier
+
 logger = create_logger()
 #logger = logging.getLogger(__name__)
