+++ flows/zeropoint.py
@@ -8,10 +8,12 @@ allows for arbitrary outlier and fitting functions.
 
 .. codeauthor:: Emir Karamehmetoglu <emir.k@phys.au.dk>
 """
+from typing import List, Optional, Dict, Union, Callable
 
 import numpy as np
 from astropy.stats import bootstrap
 from astropy.modeling import fitting
+from numpy.typing import ArrayLike
 from scipy.special import erfcinv
 
 
@@ -21,9 +23,11 @@ def sigma_from_Chauvenet(Nsamples):
     return erfcinv(1. / (2 * Nsamples)) * (2.) ** (1 / 2)
 
 
-def bootstrap_outlier(x, y, yerr, n=500, model='None', fitter='None', outlier='None', outlier_kwargs={'sigma': 3},
-                      summary='median', error='bootstrap', parnames=['intercept'], return_vals=True):
-    '''x = catalog mag, y = instrumental mag, yerr = instrumental error
+def bootstrap_outlier(x: ArrayLike, y: ArrayLike, yerr: ArrayLike, n: int = 500, model='None',
+                      fitter: Union[Callable, str] = None, outlier='None', outlier_kwargs: Optional[Dict] = None,
+                      summary: Union[Callable, str] = 'median', error: Union[Callable, str] = 'bootstrap',
+                      parnames: Optional[List] = None, return_vals: bool = True):
+    """x = catalog mag, y = instrumental mag, yerr = instrumental error
     summary = function for summary statistic, np.nanmedian by default.
     model = Linear1D
     fitter = LinearLSQFitter
@@ -31,9 +35,11 @@ def bootstrap_outlier(x, y, yerr, n=500, model='None', fitter='None', outlier='N
     outlier_kwargs, default sigma = 3
     return_vals = False will return dictionary
     Performs bootstrap with replacement and returns model.
-    '''
+    """
     summary = np.nanmedian if summary == 'median' else summary
     error = np.nanstd if error == 'bootstrap' else error
+    parnames = ['intercept'] if parnames is None else parnames
+    outlier_kwargs = {'sigma': 3} if outlier_kwargs is None else outlier_kwargs
 
     # Create index for bootstrapping
     ind = np.arange(len(x))
@@ -44,7 +50,7 @@ def bootstrap_outlier(x, y, yerr, n=500, model='None', fitter='None', outlier='N
     bootinds = bootstraps.astype(int)
 
     # Prepare fitter
-    fitter = fitting.LinearLSQFitter if fitter == "None" else fitter
+    fitter = fitting.LinearLSQFitter if fitter is None else fitter
     fitter_instance = fitting.FittingWithOutlierRemoval(fitter(), outlier, **outlier_kwargs)
     # Fit each bootstrap with model and fitter using outlier rejection at each step.
     # Then obtain summary statistic for each parameter in parnames
